refactor: organize Rust pipeline tests by method
