+++ .github/workflows/ci.yml
@@ -199,7 +199,7 @@ jobs:
       - name: Test
         # Skip some tests cause they are a space hog and run them in the separate job
         run: |
-          cargo make test -E 'package(midenc-integration-tests) - test(~rust_pipeline::examples)'
+          cargo make test -E 'package(midenc-integration-tests) - test(~rust_pipeline::(expect|functional|property)::examples)'
 
   midenc_integration_tests_examples:
     name: midenc integration tests
@@ -229,7 +229,7 @@ jobs:
       - name: Test
         # Skip some tests cause they are a space hog and run them in the separate job
         run: |
-          cargo make test -E 'test(~rust_pipeline::examples)'
+          cargo make test -E 'test(~rust_pipeline::(expect|functional|property)::examples)'
 
   cargo_miden_integration_tests:
     name: cargo-miden integration tests
