+++ tests/integration/src/rust_pipeline/compile/mod.rs
@@ -1,0 +1,1 @@
+mod sdk;
