+++ tests/integration/src/rust_pipeline/compile/sdk/base/asset.rs
