+++ tests/integration/src/rust_pipeline/compile/sdk/base/input_note.rs
