+++ tests/integration/src/rust_pipeline/compile/sdk/base/output_note.rs
