+++ tests/integration/src/rust_pipeline/compile/sdk/macros.rs
