+++ tests/integration/src/rust_pipeline/compile/sdk/mod.rs
