+++ tests/integration/src/rust_pipeline/compile/sdk/stdlib/collections.rs
