+++ tests/integration/src/rust_pipeline/compile/sdk/stdlib/hashes.rs
