+++ tests/integration/src/rust_pipeline/compile/sdk/stdlib/mod.rs
