--- tests/integration/src/rust_pipeline/examples.rs
@@ -1,385 +1,0 @@
-use std::{borrow::Borrow, collections::VecDeque};
-
-use miden_core::serde::{Deserializable, Serializable};
-use miden_debug::ToMidenRepr;
-use miden_mast_package::SectionId;
-use miden_protocol::{account::AccountComponentMetadata, note::NoteScript};
-use midenc_expect_test::{expect, expect_file};
-use midenc_frontend_wasm::WasmTranslationConfig;
-use midenc_hir::{
-    Felt, FunctionIdent, Ident, Immediate, Op, SourceSpan, SymbolTable, interner::Symbol,
-};
-use prop::test_runner::{Config, TestRunner};
-use proptest::prelude::*;
-
-use crate::{
-    CompilerTest, CompilerTestBuilder,
-    cargo_proj::project,
-    testing::{executor_with_std, stripped_mast_size_str},
-};
-
-/// Asserts that the exported procedure carrying `attribute` is unique and preserves its leaf
-/// export name.
-fn assert_unique_protocol_export(
-    package: &miden_mast_package::Package,
-    attribute: &str,
-    expected_export_name: &str,
-) {
-    let matching_exports = package
-        .mast
-        .exports()
-        .filter_map(|export| {
-            let proc_export = export.as_procedure()?;
-            proc_export.attributes.has(attribute).then_some(proc_export)
-        })
-        .collect::<Vec<_>>();
-    assert_eq!(
-        matching_exports.len(),
-        1,
-        "expected exactly one exported procedure to carry the `{attribute}` attribute",
-    );
-
-    let export_name = matching_exports[0]
-        .path
-        .last()
-        .expect("protocol export should have a procedure name");
-    assert_eq!(
-        export_name, expected_export_name,
-        "expected the `{attribute}` export to preserve the user-defined procedure name",
-    );
-}
-
-#[test]
-fn storage_example() {
-    let config = WasmTranslationConfig::default();
-    let mut test =
-        CompilerTest::rust_source_cargo_miden("../../examples/storage-example", config, []);
-
-    let package = test.compile_package();
-    let account_component_metadata_bytes = package
-        .as_ref()
-        .sections
-        .iter()
-        .find_map(|s| {
-            if s.id == SectionId::ACCOUNT_COMPONENT_METADATA {
-                Some(s.data.borrow())
-            } else {
-                None
-            }
-        })
-        .unwrap();
-    let toml = AccountComponentMetadata::read_from_bytes(account_component_metadata_bytes)
-        .unwrap()
-        .to_toml()
-        .unwrap();
-    expect![[r#"
-        name = "storage-example"
-        description = "A simple example of a Miden account storage API"
-        version = "0.1.0"
-        supported-types = ["RegularAccountUpdatableCode"]
-
-        [[storage.slots]]
-        name = "miden_storage_example::my_account::owner_public_key"
-        description = "owner public key"
-        type = "word"
-
-        [[storage.slots]]
-        name = "miden_storage_example::my_account::asset_qty_map"
-        description = "asset quantity map"
-
-        [storage.slots.type]
-        key = "word"
-        value = "felt"
-    "#]]
-    .assert_eq(&toml);
-}
-
-#[test]
-fn fibonacci() {
-    fn expected_fib(n: u32) -> u32 {
-        let mut a = 0;
-        let mut b = 1;
-        for _ in 0..n {
-            let c = a + b;
-            a = b;
-            b = c;
-        }
-        a
-    }
-
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_source_cargo_miden("../../examples/fibonacci", config, []);
-    let package = test.compile_package();
-
-    // Run the Rust and compiled MASM code against a bunch of random inputs and compare the results
-    TestRunner::default()
-        .run(&(1u32..30), move |a| {
-            let rust_out = expected_fib(a);
-            let exec = executor_with_std(vec![Felt::new(a as u64)], Some(&package));
-            let output: u32 =
-                exec.execute_into(&package.unwrap_program(), test.session.source_manager.clone());
-            dbg!(output);
-            prop_assert_eq!(rust_out, output);
-            Ok(())
-        })
-        .unwrap_or_else(|err| panic!("{err}"));
-}
-
-#[test]
-fn collatz() {
-    crate::testing::setup::enable_compiler_instrumentation();
-
-    fn expected(mut n: u32) -> u32 {
-        let mut steps = 0;
-        while n != 1 {
-            if n.is_multiple_of(2) {
-                n /= 2;
-            } else {
-                n = 3 * n + 1;
-            }
-            steps += 1;
-        }
-        steps
-    }
-
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_source_cargo_miden("../../examples/collatz", config, []);
-    let artifact_name = "collatz";
-    let package = test.compile_package();
-
-    // Run the Rust and compiled MASM code against a bunch of random inputs and compare the results
-    TestRunner::new(Config::with_cases(4))
-        .run(&(1u32..30), move |a| {
-            let rust_out = expected(a);
-            let args = a.to_felts().to_vec();
-            let exec = executor_with_std(args, Some(&package));
-            let output: u32 =
-                exec.execute_into(&package.unwrap_program(), test.session.source_manager.clone());
-            dbg!(output);
-            prop_assert_eq!(rust_out, output);
-            Ok(())
-        })
-        .unwrap_or_else(|err| {
-            panic!("{err}");
-        });
-}
-
-#[test]
-fn is_prime() {
-    crate::testing::setup::enable_compiler_instrumentation();
-
-    fn expected(n: u32) -> bool {
-        if n <= 1 {
-            return false;
-        }
-        if n <= 3 {
-            return true;
-        }
-        if n.is_multiple_of(2) || n.is_multiple_of(3) {
-            return false;
-        }
-        let mut i = 5;
-        while i * i <= n {
-            if n.is_multiple_of(i) || n.is_multiple_of(i + 2) {
-                return false;
-            }
-            i += 6;
-        }
-        true
-    }
-
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_source_cargo_miden("../../examples/is-prime", config, []);
-    let artifact_name = "is_prime";
-    let package = test.compile_package();
-    let hir = test.hir();
-
-    println!("{}", hir.borrow().as_operation());
-
-    // Run the Rust and compiled MASM code against a bunch of random inputs and compare the results
-    TestRunner::new(Config::with_cases(100))
-        .run(&(1u32..30), move |a| {
-            let rust_out = expected(a);
-
-            // Test the IR
-            let mut evaluator =
-                midenc_hir_eval::HirEvaluator::new(hir.borrow().as_operation().context_rc());
-            let op = hir
-                .borrow()
-                .symbol_manager()
-                .lookup_symbol_ref(
-                    &midenc_hir::SymbolPath::new([
-                        midenc_hir::SymbolNameComponent::Component("is_prime".into()),
-                        midenc_hir::SymbolNameComponent::Leaf("entrypoint".into()),
-                    ])
-                    .unwrap(),
-                )
-                .unwrap();
-            let result = evaluator
-                .eval(&op.borrow(), [midenc_hir_eval::Value::Immediate((a as i32).into())])
-                .unwrap_or_else(|err| panic!("{err}"));
-            let midenc_hir_eval::Value::Immediate(Immediate::I32(result)) = result[0] else {
-                //return Err(TestCaseError::fail(format!(
-                panic!("expected i32 immediate for input {a}, got {:?}", result[0]);
-                //)));
-            };
-            prop_assert_eq!(rust_out as i32, result);
-
-            let args = a.to_felts().to_vec();
-            let exec = executor_with_std(args, Some(&package));
-            let output: u32 =
-                exec.execute_into(&package.unwrap_program(), test.session.source_manager.clone());
-            dbg!(output);
-            prop_assert_eq!(rust_out as u32, output);
-            Ok(())
-        })
-        .unwrap_or_else(|err| {
-            panic!("{err}");
-        });
-}
-
-#[test]
-fn counter_contract() {
-    let config = WasmTranslationConfig::default();
-    let mut builder_release = CompilerTestBuilder::rust_source_cargo_miden(
-        "../../examples/counter-contract",
-        config.clone(),
-        [],
-    );
-    builder_release.with_release(true);
-    let mut test_release = builder_release.build();
-    let package = test_release.compile_package();
-    let account_component_metadata_bytes = package
-        .as_ref()
-        .sections
-        .iter()
-        .find_map(|s| {
-            if s.id == SectionId::ACCOUNT_COMPONENT_METADATA {
-                Some(s.data.borrow())
-            } else {
-                None
-            }
-        })
-        .unwrap();
-    let toml = AccountComponentMetadata::read_from_bytes(account_component_metadata_bytes)
-        .unwrap()
-        .to_toml()
-        .unwrap();
-    expect![[r#"
-        name = "counter-contract"
-        description = "A simple example of a Miden counter contract using the Account Storage API"
-        version = "0.1.0"
-        supported-types = ["RegularAccountUpdatableCode"]
-
-        [[storage.slots]]
-        name = "miden_counter_contract::counter_contract::count_map"
-        description = "counter contract storage map"
-
-        [storage.slots.type]
-        key = "word"
-        value = "felt"
-    "#]]
-    .assert_eq(&toml);
-}
-
-#[test]
-fn counter_contract_debug_build() {
-    // This build checks the dev profile build compilation for counter-contract
-    // see https://github.com/0xMiden/compiler/issues/510
-    let config = WasmTranslationConfig::default();
-    let mut builder =
-        CompilerTestBuilder::rust_source_cargo_miden("../../examples/counter-contract", config, []);
-    builder.with_release(false);
-    let mut test = builder.build();
-    let package = test.compile_package();
-}
-
-#[test]
-fn counter_note() {
-    // build and check counter-note
-    let config = WasmTranslationConfig::default();
-    let builder =
-        CompilerTestBuilder::rust_source_cargo_miden("../../examples/counter-note", config, []);
-
-    let mut test = builder.build();
-
-    let package = test.compile_package();
-    assert!(package.is_library(), "expected library");
-    let _note_script =
-        NoteScript::from_package(package.as_ref()).expect("expected a note-script package");
-    assert_unique_protocol_export(package.as_ref(), "note_script", "run");
-
-    // TODO: uncomment after the testing environment implemented (node, devnet, etc.)
-    //
-    // let mut exec = Executor::new(vec![]);
-    // for dep_path in test.dependencies {
-    //     let account_package =
-    //         Arc::new(Package::read_from_bytes(&std::fs::read(dep_path).unwrap()).unwrap());
-    //     exec.dependency_resolver_mut()
-    //         .add(account_package.digest(), account_package.into());
-    // }
-    // exec.with_dependencies(&package.manifest.dependencies).unwrap();
-    // let trace = exec.execute(&package.unwrap_program(), &test.session);
-}
-
-#[test]
-fn basic_wallet_and_p2id() {
-    let config = WasmTranslationConfig::default();
-    let mut account_test =
-        CompilerTest::rust_source_cargo_miden("../../examples/basic-wallet", config.clone(), []);
-    let account_package = account_test.compile_package();
-    assert!(account_package.is_library(), "expected library");
-    expect!["35906"].assert_eq(stripped_mast_size_str(&account_package));
-
-    let mut tx_script_test = CompilerTest::rust_source_cargo_miden(
-        "../../examples/basic-wallet-tx-script",
-        config.clone(),
-        [],
-    );
-    let tx_script_package = tx_script_test.compile_package();
-    assert!(tx_script_package.is_program(), "expected program");
-    expect!["56437"].assert_eq(stripped_mast_size_str(&tx_script_package));
-
-    let mut p2id_test =
-        CompilerTest::rust_source_cargo_miden("../../examples/p2id-note", config.clone(), []);
-    let note_package = p2id_test.compile_package();
-    assert!(note_package.is_library(), "expected library");
-    expect!["53082"].assert_eq(stripped_mast_size_str(&note_package));
-
-    let mut p2ide_test =
-        CompilerTest::rust_source_cargo_miden("../../examples/p2ide-note", config, []);
-    let p2ide_package = p2ide_test.compile_package();
-    assert!(p2ide_package.is_library(), "expected library");
-    expect!["62672"].assert_eq(stripped_mast_size_str(&p2ide_package));
-}
-
-#[test]
-fn auth_component_no_auth() {
-    let config = WasmTranslationConfig::default();
-    let mut test =
-        CompilerTest::rust_source_cargo_miden("../../examples/auth-component-no-auth", config, []);
-    let auth_comp_package = test.compile_package();
-    assert!(auth_comp_package.is_library());
-    assert_unique_protocol_export(auth_comp_package.as_ref(), "auth_script", "auth-procedure");
-
-    // Test that the package loads
-    let bytes = auth_comp_package.to_bytes();
-    let _loaded_package = miden_mast_package::Package::read_from_bytes(&bytes).unwrap();
-}
-
-#[test]
-fn auth_component_rpo_falcon512() {
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_source_cargo_miden(
-        "../../examples/auth-component-rpo-falcon512",
-        config,
-        [],
-    );
-    let auth_comp_package = test.compile_package();
-    assert!(auth_comp_package.is_library());
-    assert_unique_protocol_export(auth_comp_package.as_ref(), "auth_script", "check-signature");
-
-    // Test that the package loads
-    let bytes = auth_comp_package.to_bytes();
-    let _loaded_package = miden_mast_package::Package::read_from_bytes(&bytes).unwrap();
-}
