+++ tests/integration/src/rust_pipeline/expect/mod.rs
@@ -1,0 +1,1 @@
+mod examples;
