+++ tests/integration/src/rust_pipeline/functional/abi_transform/mod.rs
@@ -1,0 +1,1 @@
+mod tx_kernel;
