+++ tests/integration/src/rust_pipeline/functional/apps.rs
