+++ tests/integration/src/rust_pipeline/functional/debug_source_locations.rs
