+++ tests/integration/src/rust_pipeline/functional/memory/mod.rs
@@ -1,0 +1,2 @@
+mod copy;
+mod set;
