+++ tests/integration/src/rust_pipeline/functional/mod.rs
@@ -1,0 +1,7 @@
+mod abi_transform;
+mod apps;
+mod debug_source_locations;
+mod examples;
+mod memory;
+mod regressions;
+mod types;
