+++ tests/integration/src/rust_pipeline/functional/regressions.rs
