+++ tests/integration/src/rust_pipeline/functional/types/mod.rs
