+++ tests/integration/src/rust_pipeline/functional/types/types_src/enum.rs
