+++ tests/integration/src/rust_pipeline/functional/types/types_src/static_mut.rs
