+++ tests/integration/src/rust_pipeline/property/abi_transform/mod.rs
@@ -1,3 +1,2 @@
 mod advice_map;
 mod stdlib;
-mod tx_kernel;
