+++ tests/integration/src/rust_pipeline/property/arithmetic.rs
@@ -11,9 +11,9 @@ use proptest::{
     test_runner::{TestError, TestRunner},
 };
 
-use super::run_masm_vs_rust;
 use crate::{
     CompilerTest,
+    rust_pipeline::run_masm_vs_rust,
     testing::{Initializer, eval_package, setup},
 };
 
@@ -1037,619 +1037,3 @@ test_unary_op_total!(bnot, !, u32);
 test_unary_op_total!(bnot, !, u16);
 test_unary_op_total!(bnot, !, u8);
 test_unary_op_total!(bnot, !, bool);
-
-#[test]
-fn test_hmerge() {
-    let main_fn = r#"
-	        (f0: miden_stdlib_sys::Felt, f1: miden_stdlib_sys::Felt, f2: miden_stdlib_sys::Felt, f3: miden_stdlib_sys::Felt, f4: miden_stdlib_sys::Felt, f5: miden_stdlib_sys::Felt, f6: miden_stdlib_sys::Felt, f7: miden_stdlib_sys::Felt) -> miden_stdlib_sys::Felt {
-	            let digest1 = miden_stdlib_sys::Digest::new([f0, f1, f2, f3]);
-	            let digest2 = miden_stdlib_sys::Digest::new([f4, f5, f6, f7]);
-	            let digests = [digest1, digest2];
-	            let res = miden_stdlib_sys::intrinsics::crypto::merge(digests);
-	            res.inner[0]
-	        }"#
-	        .to_string();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys("hmerge", &main_fn, config, []);
-
-    let package = test.compile_package();
-
-    // Run the Rust and compiled MASM code against a bunch of random inputs and compare the results
-    let config = proptest::test_runner::Config::with_cases(16);
-    let res = TestRunner::new(config).run(
-        &any::<([miden_debug::Felt; 4], [miden_debug::Felt; 4])>(),
-        move |(felts_in1, felts_in2)| {
-            let raw_felts_in1: [Felt; 4] = [
-                felts_in1[0].into(),
-                felts_in1[1].into(),
-                felts_in1[2].into(),
-                felts_in1[3].into(),
-            ];
-
-            let raw_felts_in2: [Felt; 4] = [
-                felts_in2[0].into(),
-                felts_in2[1].into(),
-                felts_in2[2].into(),
-                felts_in2[3].into(),
-            ];
-            let digests_in =
-                [miden_core::Word::from(raw_felts_in1), miden_core::Word::from(raw_felts_in2)];
-            let digest_out = miden_core::crypto::hash::Poseidon2::merge(&digests_in);
-
-            let felts_out: [miden_debug::Felt; 4] = [
-                miden_debug::Felt(digest_out[0]),
-                miden_debug::Felt(digest_out[1]),
-                miden_debug::Felt(digest_out[2]),
-                miden_debug::Felt(digest_out[3]),
-            ];
-
-            let args = [
-                raw_felts_in1[0],
-                raw_felts_in1[1],
-                raw_felts_in1[2],
-                raw_felts_in1[3],
-                raw_felts_in2[0],
-                raw_felts_in2[1],
-                raw_felts_in2[2],
-                raw_felts_in2[3],
-            ];
-            eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-                let res: Felt = trace.parse_result().unwrap();
-                prop_assert_eq!(res, digest_out[0]);
-                Ok(())
-            })?;
-
-            Ok(())
-        },
-    );
-
-    match res {
-        Err(TestError::Fail(_, value)) => {
-            panic!("Found minimal(shrinked) failing case: {value:?}");
-        }
-        Ok(_) => (),
-        _ => panic!("Unexpected test result: {res:?}"),
-    }
-}
-
-#[test]
-fn test_memory_copy_aligned_bytes() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u32; 12], src: &[u32; 16]) {
-            unsafe {
-                let src_ptr = (src.as_ptr() as *const u8).add(4);
-                let dst_ptr = dst.as_mut_ptr() as *mut u8;
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 48);
-            }
-        }
-
-        let mut src = [0u32; 16];
-        let src_bytes = src.as_mut_ptr() as *mut u8;
-        let mut i = 0usize;
-        while i < 64 {
-            unsafe { *src_bytes.add(i) = i as u8; }
-            i += 1;
-        }
-
-        let mut dst = [0u32; 12];
-        do_copy(&mut dst, &src);
-
-        let dst_bytes = dst.as_ptr() as *const u8;
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 48 {
-            let observed = unsafe { *dst_bytes.add(i) };
-            if observed != (i as u8).wrapping_add(4) {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_aligned_bytes_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_u128_fast_path() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u128; 2], src: &[u128; 3]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(1);
-                let dst_ptr = dst.as_mut_ptr();
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 2);
-            }
-        }
-
-        let src = [
-            0x00112233445566778899aabbccddeeff_u128,
-            0x102132435465768798a9bacbdcedfe0f_u128,
-            0xfedcba98765432100123456789abcdef_u128,
-        ];
-        let mut dst = [0u128; 2];
-        do_copy(&mut dst, &src);
-
-        let expected = [src[1], src[2]];
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 2 {
-            if dst[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_u128_fast_path",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_multiword_fast_path() {
-    let main_fn = r#"() -> Felt {
-        struct Chunk([u128; 2]);
-
-        #[inline(never)]
-        fn do_copy(dst: &mut [Chunk; 1], src: &[Chunk; 2]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(1);
-                let dst_ptr = dst.as_mut_ptr();
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 1);
-            }
-        }
-
-        let src = [
-            Chunk([
-                0x00112233445566778899aabbccddeeff_u128,
-                0x112233445566778899aabbccddeeff00_u128,
-            ]),
-            Chunk([
-                0xaabbccddeeff00112233445566778899_u128,
-                0xffeeddccbbaa99887766554433221100_u128,
-            ]),
-        ];
-        let mut dst = [Chunk([0u128; 2])];
-        do_copy(&mut dst, &src);
-
-        let expected = &src[1].0;
-        let observed = &dst[0].0;
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 2 {
-            if observed[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_multiword_fast_path",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_aligned_addresses_misaligned_count() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u32; 12], src: &[u32; 16]) {
-            unsafe {
-                let src_ptr = (src.as_ptr() as *const u8).add(4);
-                let dst_ptr = dst.as_mut_ptr() as *mut u8;
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 47);
-            }
-        }
-
-        let mut src = [0u32; 16];
-        let src_bytes = src.as_mut_ptr() as *mut u8;
-        let mut i = 0usize;
-        while i < 64 {
-            unsafe { *src_bytes.add(i) = i as u8; }
-            i += 1;
-        }
-
-        let mut dst = [0xffff_ffffu32; 12];
-        do_copy(&mut dst, &src);
-
-        let dst_bytes = dst.as_ptr() as *const u8;
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 48 {
-            let observed = unsafe { *dst_bytes.add(i) };
-            let expected = if i < 47 {
-                (i as u8).wrapping_add(4)
-            } else {
-                0xff
-            };
-            if observed != expected {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_aligned_addresses_misaligned_count_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_unaligned() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u8; 48], src: &[u8; 64]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(3);
-                let dst_ptr = dst.as_mut_ptr();
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 48);
-            }
-        }
-
-        let mut src = [0u8; 64];
-        let mut i = 0usize;
-        while i < 64 {
-            src[i] = i as u8;
-            i += 1;
-        }
-
-        let mut dst = [0u8; 48];
-        do_copy(&mut dst, &src);
-
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 48 {
-            if dst[i] != (i as u8).wrapping_add(3) {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_unaligned_src_len_48_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_unaligned_dst() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u8; 53], src: &[u8; 64]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(3);
-                let dst_ptr = dst.as_mut_ptr().add(5);
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 48);
-            }
-        }
-
-        let mut src = [0u8; 64];
-        let mut i = 0usize;
-        while i < 64 {
-            src[i] = i as u8;
-            i += 1;
-        }
-
-        let mut dst = [0xffu8; 53];
-        do_copy(&mut dst, &src);
-
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 53 {
-            let expected = if i < 5 { 0xff } else { (i as u8).wrapping_sub(2) };
-            if dst[i] != expected {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_unaligned_dst_len_48_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_unaligned_dst_short_count() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u8; 8], src: &[u8; 16]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(3);
-                let dst_ptr = dst.as_mut_ptr().add(2);
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 3);
-            }
-        }
-
-        let mut src = [0u8; 16];
-        let mut i = 0usize;
-        while i < 16 {
-            src[i] = i as u8;
-            i += 1;
-        }
-
-        let mut dst = [0xffu8; 8];
-        do_copy(&mut dst, &src);
-
-        let expected = [0xffu8, 0xff, 3, 4, 5, 0xff, 0xff, 0xff];
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 8 {
-            if dst[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_unaligned_dst_short_count_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_copy_unaligned_zero_count() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_copy(dst: &mut [u8; 8], src: &[u8; 16]) {
-            unsafe {
-                let src_ptr = src.as_ptr().add(1);
-                let dst_ptr = dst.as_mut_ptr().add(2);
-                core::ptr::copy_nonoverlapping(src_ptr, dst_ptr, 0);
-            }
-        }
-
-        let mut src = [0u8; 16];
-        let mut i = 0usize;
-        while i < 16 {
-            src[i] = i as u8;
-            i += 1;
-        }
-
-        let mut dst = [0xffu8; 8];
-        do_copy(&mut dst, &src);
-
-        let expected = [0xffu8; 8];
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 8 {
-            if dst[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_copy_unaligned_zero_count_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_set_unaligned() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_set(dst: &mut [u8; 11]) {
-            unsafe {
-                let dst_ptr = dst.as_mut_ptr().add(3);
-                core::ptr::write_bytes(dst_ptr, 0x5a, 5);
-            }
-        }
-
-        let mut dst = [0xffu8; 11];
-        do_set(&mut dst);
-
-        let expected = [0xffu8, 0xff, 0xff, 0x5a, 0x5a, 0x5a, 0x5a, 0x5a, 0xff, 0xff, 0xff];
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 11 {
-            if dst[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test =
-        CompilerTest::rust_fn_body_with_stdlib_sys("memory_set_unaligned_u8s", main_fn, config, []);
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[test]
-fn test_memory_set_unaligned_zero_count() {
-    let main_fn = r#"() -> Felt {
-        #[inline(never)]
-        fn do_set(dst: &mut [u8; 11]) {
-            unsafe {
-                let dst_ptr = dst.as_mut_ptr().add(3);
-                core::ptr::write_bytes(dst_ptr, 0x5a, 0);
-            }
-        }
-
-        let mut dst = [0xffu8; 11];
-        do_set(&mut dst);
-
-        let expected = [0xffu8; 11];
-        let mut mismatches = 0u32;
-        let mut i = 0usize;
-        while i < 11 {
-            if dst[i] != expected[i] {
-                mismatches += 1;
-            }
-            i += 1;
-        }
-
-        Felt::from_u32(mismatches)
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "memory_set_unaligned_zero_count_u8s",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, Felt::ZERO);
-        Ok(())
-    })
-    .unwrap();
-}
