+++ tests/integration/src/rust_pipeline/property/intrinsics.rs
