+++ tests/integration/src/rust_pipeline/property/mod.rs
@@ -1,0 +1,5 @@
+mod abi_transform;
+mod arithmetic;
+mod crypto;
+mod examples;
+mod intrinsics;
